+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -102,6 +102,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Shader framework refactor: Changed ClampRoughness signature to include PreLightData access.
 - HDRPWizard window is now in Window > General > HD Render Pipeline Wizard
 - Moved StaticLightingSky to LightingWindow
+- Removes the current "Scene Settings" and replace them with "Sky & Fog Settings" (with Physically Based Sky and Volumetric Fog).
 
 ## [6.7.0-preview] - 2019-05-16